Fix building on LXD hosts.
There's a bug with the value of $SNAPCRAFT_PART_BUILD in Multipass vs in LXD.
In LXD, the value does not contain source_subdir and in Multipass it does.
Since going forward I'd like to use the Launchpad infrastructure to build,
this commit fixes building in LXD at the expense of Multipass.